refactor source code
