+++ chloe/gameoflife/src/main/java/com/katas/gameoflife/CellState.java
