+++ chloe/gameoflife/src/main/java/com/katas/gameoflife/GameOfLife.java
