+++ chloe/minesweeper/src/main/java/com/katas/minesweeper/IArea.java
