+++ chloe/minesweeper/src/main/java/com/katas/minesweeper/Square.java
@@ -1,4 +1,4 @@
-package com.katas.minesweeper.squares;
+package com.katas.minesweeper;
 
 /**
  * @author : Chloé Mahalin
