+++ chloe/minesweeper/src/main/java/com/katas/minesweeper/SquareArea.java
