+++ chloe/rpn/src/main/java/com/katas/rpn/stack/operators/IOperator.java
